requirements.txt
bscscan-python==1.0.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
urllib3==1.26.4

pythonpancakes~=1.0.0
nose~=1.3.7